+++ ChangeLog.md
@@ -2,6 +2,7 @@
 <summary> 3.14.1 - 2026-06-05 </summary>
 
 - Added configuration setting to set the initial charge when acquiring night vision.
+- Added configuration setting to disable Locksmith minigame trigger when colliding with locked doors, giving the only option of interacting with them to start it.
 - Fixed info display for TZP Buffer showing incorrect stats on the first level.
 - Fixed info display for Bullet Resistance for missing a percentage sign.